improvement: 优化用户角色管理

+++ src/main/java/com/xxx/aimianshi/userrole/service/impl/UserRoleServiceImpl.java
@@ -1,11 +1,14 @@
 package com.xxx.aimianshi.userrole.service.impl;
 
+import com.xxx.aimianshi.common.exception.BizException;
+import com.xxx.aimianshi.common.utils.ThrowUtils;
 import com.xxx.aimianshi.role.domain.entity.Role;
 import com.xxx.aimianshi.role.enums.RoleEnum;
 import com.xxx.aimianshi.role.repository.RoleRepository;
 import com.xxx.aimianshi.userrole.domain.entity.UserRole;
 import com.xxx.aimianshi.userrole.repository.UserRoleRepository;
 import com.xxx.aimianshi.userrole.service.UserRoleService;
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -21,24 +24,34 @@ public class UserRoleServiceImpl implements UserRoleService {
     }
 
     @Override
-    public boolean addUserRole(Long userId, RoleEnum roleEnum) {
+    public void addUserRole(Long userId, RoleEnum roleEnum) {
         Role role = roleRepository.findByRoleName(roleEnum.getName())
                 .orElseThrow(() -> new RuntimeException("Role not found"));
-        return userRoleRepository.save(new UserRole(null, userId, role.getId()));
+        try {
+            userRoleRepository.save(new UserRole(null, userId, role.getId()));
+        } catch (DuplicateKeyException e) {
+            throw new BizException("user already has role: " + roleEnum.getName());
+        }
     }
 
     /**
      * 清理用户所有角色
      */
     @Override
-    public boolean deleteUserRoles(Long userId) {
+    public void deleteUserRoles(Long userId) {
         List<UserRole> userRoles = userRoleRepository.getByUserId(userId);
-        return userRoleRepository.removeByIds(userRoles);
+        if (userRoles.isEmpty()) {
+            return;
+        }
+        userRoleRepository.removeByIds(userRoles);
     }
 
     @Override
-    public void add(UserRole userRole) {
-        userRoleRepository.save(userRole);
+    public void deleteUserRole(Long userId, RoleEnum roleEnum) {
+        UserRole userRole = userRoleRepository.getOptByUserIdAndRole(userId, roleEnum.getName())
+                .orElseThrow(() -> new RuntimeException("user role not found"));
+        boolean remove = userRoleRepository.removeById(userRole);
+        ThrowUtils.throwIf(!remove, "delete failed, maybe the role does not exist");
     }
 
 }
